% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RAGP_norm}
\alias{RAGP_norm}
\title{Normalized HT-qPCR data for single neurons from the Right Atrial Ganglionated Plexus of a Yucatan Minipig}
\format{
A matrix with 174 rows (genes) and 405 columns (samples):
}
\source{
Data can be found at GEO reference ID: GSE149212 and on sparc.science portal at DOI: 10.26275/5jki-b4er
}
\usage{
RAGP_norm
}
\description{
HT-qPCR data for 174 genes for 405 single neurons across 4 Pigs. These are the genes that passed QC
for all 4 animals. See \link[dataVisEasy]{RAGP_rawCt} for the raw Ct values and \link[dataVisEasy]{RAGP_annots}
for annotations for all samples. This data is used in examples for functions in the dataVisEasy package.
}
\keyword{datasets}
